Switch back to building both CommonJS and ES6
This means that npm link doesn't appear to work as when resolving linked modules it appears that our current setup for apps uses the CommonJS version which then only resolves dependencies relative to the linked module.

This switches back to outputting both CommonJS and ES6 so that Jest (which doesn't support ES6 out of the box) can use the CommonJS version of the module.