Add PromptOAuth component
This allows for the UI of a tool to continue to be shown, but for a prompt to grant access to be overlaid on-top.

This was added as a new component as it's behaviour is quite different to the existing OAuth component in that it doesn't have any child components.