Force a resize when limiting (or unlimiting).
When limiting the height (for example because we are displaying a modal) we don't force the new size to be sent through the postMessage, but instead rely on the content changing to cause an update. But it's not always true that the content will changes (and trigger the mutation observer) so instead we always force a resize when changing the limit.

This also updates the component so that it only pushes a resize if it's different to the last size we had.