Accept server value directly
Due to this component being extracted from an application the `server` prop is expected to be a object with a `proxyServer` property, this isn't needed and is confusing to consumers of the application.

This adds support for the `server` prop to be specified directly and makes the component simpler to user. We still support the old format.